Food Tracker - update logic

+++ app/src/main/java/com/example/boilerplate/main/food_tracker/model/Food.kt
@@ -2,6 +2,6 @@ package com.example.boilerplate.main.food_tracker.model
 
 data class Food(
     val name: String,
-    val calories: Int,
+    val calories: Long,
     val type: String
 )
